codex-bravo: GLM-5.2 PR #13 residual defect audit + 4/4 Defect-2 coverage
Merged by ar-trn/validator per operator directive 2026-08-28T03:00Z. Strict superset of closed AR PR #2. Covers all 3 PR #13 defects at 4/4 sites + tests + GATE-STILL-FAILING receipt. 10-tok gate re-run pending GLM-5.2-FP8 checkpoint pull.

+++ vllm_neuron/model/glm52_moe_dsa/indexer.py
@@ -212,10 +212,11 @@ class Glm52FullIndexer(nn.Module):
             # At the 2,048-token baseline, GLM's index_topk equals the full
             # context. Selecting every position does not require ranking them.
             # Avoid torch.topk(k == context), whose full-sort HLO is unsupported
-            # on Trn2, and preserve the same complete index set in logical order.
+            # on Trn2, preserve logical order, and keep the signed-int64 index
+            # invariant required by the PR #13 sentinel-wrap fix.
             indices = torch.arange(
                 selected,
-                dtype=torch.int32,
+                dtype=torch.int64,
                 device=scores.device,
             )
             leading_ones = (1,) * (scores.ndim - 1)
